Refactoring the line parsing
